+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added support for webhookForSlotFilling (dialogflow)
+- Added support for webhookUsed (dialogflow)
+- Added suppport for intent responses (dialogflow)
 
 ### Fixed